% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupytext.R
\name{add_jupytext}
\alias{add_jupytext}
\title{Insert jupytext configuration to R Markdown}
\usage{
add_jupytext(path)
}
\arguments{
\item{path}{path to R Markdown file}
}
\description{
Insert jupytext configuration to R Markdown
}
