% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifrost-package.R
\docType{package}
\name{bifrost-package}
\alias{bifrost}
\alias{bifrost-package}
\title{bifrost: Capelin assessment}
\description{
Capelin stock assessment software has historically been distributed over
    many different platforms. This package seeks to unify and gather all in one place. 
    One should be able to execute everything in the current assessment framework
    from this platform..
}
\keyword{internal}
