% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{calcSSB}
\alias{calcSSB}
\title{Calculate SSB from fit (all ages)}
\usage{
calcSSB(fit, perage = FALSE)
}
\arguments{
\item{fit}{maturity object,}

\item{perage}{logical, should the SSB be reported by age?}
}
\value{
SSB, either as numeric vector per age or total given by perage argument.
}
\description{
Calculate SSB from fit (all ages)
}
\examples{
\dontrun{
calcSSB(fit, perage = TRUE)
}

}
