% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateAllDeltaGIC}
\alias{calculateAllDeltaGIC}
\title{Calculate Delta GIC Scores Relative to a Baseline Model}
\usage{
calculateAllDeltaGIC(model_results, painted_tree_list)
}
\description{
This function computes the difference in Generalized Information Criterion (GIC) scores
between a baseline model (assumed to be the first in the list) and a series of alternative
models fitted to painted phylogenetic trees. It returns a named vector of delta GIC values,
representing the improvement or decline in model fit relative to the baseline.
}
