% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captool_simulation.R
\name{captool}
\alias{captool}
\title{Function for running CAPTOOL projection}
\usage{
captool(data_list, nsim = 50000, cap_cv = 0.2, cod_cv = 0.3, plot = TRUE)
}
\arguments{
\item{data_list}{list of data input (see details)}

\item{nsim}{number of simulations}

\item{cap_cv}{capelin coefficient of variation (cv)}

\item{cod_cv}{cod coefficient of variation (cv)}

\item{plot}{boolean, should the projection be plotted?}
}
\value{
list of output from the projection.
}
\description{
Provide a data_list of inputs and run the projection from October 1st to April 1st.
Should reproduce the old captool (excel-based) software.
}
\details{
The 'data_list' should contain year, cap, catches, cod0, cod1, svalbard,
stochasticHistory and (optionally) captool. The year is the assessment year
(integer), cap is a table containing numbers at age (0-5) and length (32
length groups). Catches is a vector giving the catches in jan, feb and march.
The data frames cod0 and cod1 is the cod data for year Y and year Y+1
(projection) from the cod assessment. The data frame 'svalbard' contains the
historical svalbard porportions from the different years we are to sample
from divided by cod age. The data frame 'stochasticHistory' contains
historical samples of parameters we are to sample from, i.e. p1, p2, p3 (p3 =
M = natural mortality) and Cmax and Chalf. Lastly, if comparion to
captool (the excel-based software) is to be made, one can also provide the
projected quantiles from captool as an argument at these will be added to the
figure showing the projection.
}
\examples{
# captool(data_list)

}
