% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{captoolSim}
\alias{captoolSim}
\title{Captool simulations}
\usage{
captoolSim(fit, p = c(3.5, 13.87, 0.1), nsim = 15000, cv = 0.2, useEst = FALSE)
}
\arguments{
\item{fit}{maturity object}

\item{p}{parameter vector of length 3, p1, p2 and p3. Defaults to values used in 2011.}

\item{nsim}{number of simulations}

\item{cv}{numeric, coefficient of variation.}

\item{useEst}{If TRUE, the parameters from fit is used. If FALSE the specified p-vector is used.}
}
\value{
list containing simulations and quantiles (50%, 75%, 95%, 25% and 5%).
}
\description{
Captool simulations
}
\examples{
\dontrun{
captoolSim(mFit, nsim = 15000)
}

}
