% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\docType{data}
\name{catch}
\alias{catch}
\title{Catch data for capelin}
\format{
A data frame with 240 rows and 14 variables:
\describe{
  \item{age}{integer, age of fish}
  \item{year}{integer, year the fish was caught}
  \item{winter05}{numeric, abundance in winter x 0.5}
  \item{winter03}{numeric, abundance in winter x 0.3}
  \item{winter02}{numeric, abundance in winter x 0.2}
  \item{spring01}{numeric, abundance in spring x 0.1}
  \item{spring05}{numeric, abundance in spring x 0.5}
  \item{spring04}{numeric, abundance in spring x 0.4}
  \item{fill0}{numeric, 0, no fish caught in summer months}
  \item{fill1}{numeric, 0, no fish caught in summer months}
  \item{fill2}{numeric, 0, no fish caught in summer months}
  \item{fill3}{numeric, 0, no fish caught in summer months}
  \item{autumn08}{numeric, abundance in autumn x 0.8}
  \item{autumn02}{numeric, abundance in autumn x 0.2}
}
}
\source{
\url{https://www.imr.no}
}
\usage{
catch
}
\description{
A dataset containing the number of capelin per age, year and length group
}
\keyword{datasets}
