% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumptionData.R
\docType{data}
\name{consumptionData}
\alias{consumptionData}
\title{Consumption data list: consumptionData}
\format{
A list containting the following elements:
\describe{
  \item{Cco}{Empirical cod consumption of capelin by year and cod age }
  \item{Nco}{Number of cod by year and age (in 10 billions)}
  \item{Oco}{Proportion of mature cod by year and age}
  \item{SV}{Proportion of cod in Svalbard region}
  \item{Wco}{Mean weight of cod by year and age (in grams)}
  \item{Fco}{Fishing mortiality of cod by year and age}
  \item{Mco}{Natural mortality of cod by year and age}
  \item{Nmc}{Number of capelin by year and age (in 10 billions)}
  \item{Wmc}{Mean weight of capelin by year and age (in grams)}
  \item{p3}{Total moratility rate of capelin from 1Oct - 1Nov by year}
}
}
\source{
\url{https://www.imr.no}
}
\usage{
consumptionData
}
\description{
A data frame of data used to estimate consumption model for the years 1984-2010,
for cod ages 1-10 and capelin ages 1-5.
}
\keyword{datasets}
