% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMaturityData.R
\name{createMaturityData}
\alias{createMaturityData}
\title{Title}
\usage{
createMaturityData(cap, catch, min_age = 3, max_age = 4, start_year, end_year)
}
\arguments{
\item{cap}{data.frame, cap table. Should contain the following column names:
meanlength(cm), 1,2,3,4,5 and year, where 1-4 are ages and 5 is a 5+ group.}

\item{catch}{data.frame, catches data. Should contain 14 columns including
age, year and the remaining 12 for each month.}

\item{min_age}{integer. Minimum age}

\item{max_age}{integer. Maximum age}

\item{start_year}{integer. Start year}

\item{end_year}{integer. End year}
}
\value{
list, data argument for TMB
}
\description{
Title
}
\examples{
data(cap)
data(catch)
data <- createMaturityData(
cap = cap,
catch = catch,
min_age = 3,
max_age = 4,
start_year = 2017,
end_year = 2018)
str(data)


}
