% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMaturityParameter.R
\name{createMaturityParameters}
\alias{createMaturityParameters}
\title{Set up parameter list for maturity}
\usage{
createMaturityParameters(parameter, year)
}
\arguments{
\item{parameter}{data.frame, initial values for p1,p2,p3 and nu by year}

\item{year}{integer, year}
}
\value{
list, parameter object for running maturity estimation by TMB
}
\description{
Set up parameter list for maturity
}
\examples{
data(maturityInitialParameters)
createMaturityParameters(parameter=maturityInitialParameters, year = 2017)
}
