% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateConsumption.R
\name{estimateConsumption}
\alias{estimateConsumption}
\title{Function for estimating consumption model}
\usage{
estimateConsumption(data, parameters, ...)
}
\arguments{
\item{data}{list of data inputs}

\item{parameters}{list of parameters}

\item{...}{additional arguments for TMB::MakeADFun}
}
\value{
list of TMB object and optimization object
}
\description{
Function for estimating consumption model
}
\examples{
\dontrun{
data("consumptionData")
par.list <- list(
  logCmax   = log(1.2),
  logChalf  = log(1e2),
  logalpha  = log(2),
  logbeta   = log(2),
  logSigma  = log(1e3)
)
cFit <- estimateConsumption(data = consumptionData, parameters = par.list, silent =TRUE,
                            map  = list(logalpha = factor(NA),
                                        logbeta = factor(NA)))
}
}
