% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaturity.R
\name{estimateMaturity}
\alias{estimateMaturity}
\title{Function for estimating maturity}
\usage{
estimateMaturity(data, parameters, ...)
}
\arguments{
\item{data}{list of data inputs}

\item{parameters}{list of parameters}

\item{...}{additional arguments for TMB::MakeADFun}
}
\value{
list of TMB object and optimization object
}
\description{
Function for estimating maturity
}
\examples{
\dontrun{
data(cap)
data(catch)
data(maturityInitialParameters)

#..set up data list..
data.list <- createMaturityData(cap,
                                catch,
                                min_age = 2,
                                max_age = 3,
                                start_year =1972,
                                end_year = 2010)
# ..set up parameter list..
par.list <- createMaturityParameters(parameter = maturityInitialParameters,
                                    year = data.list$start_year, agegr = "2-3")
mFit <- estimateMaturity(data = data.list, parameters = par.list, silent =TRUE)
}
}
