% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captool_simulation.R
\name{grid.search.catch}
\alias{grid.search.catch}
\title{Function for setting catch, when zero catch gives prediction above blim}
\usage{
grid.search.catch(
  captool_run,
  catchgrid = seq(0, 0.1, 0.02),
  catch.distribution = c(0, 0.3, 0.7),
  blim = 0.2,
  seed = 1,
  ...
)
}
\arguments{
\item{captool_run}{output from captool function}

\item{catchgrid}{grid of potential catch quotas}

\item{catch.distribution}{vector of length 3 given ratio of catch in jan, feb and mar}

\item{blim}{Blim}

\item{seed}{integer, for not changing the random input}

\item{...}{}
}
\value{
captool object
}
\description{
Function for setting catch, when zero catch gives prediction above blim
}
