% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaturity.R
\name{maturing}
\alias{maturing}
\title{Function for calculating maturity from parameter estimates}
\usage{
maturing(meanlength, p1, p2)
}
\arguments{
\item{meanlength}{numeric, mean length per length group}

\item{p1}{numeric, p1 estimate}

\item{p2}{numeric, p2 estimate}
}
\value{
maturity at length
}
\description{
Function for calculating maturity from parameter estimates
}
\examples{
maturing(6, 1, 14)
}
