% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturityInitialParameters.R
\docType{data}
\name{maturityInitialParameters}
\alias{maturityInitialParameters}
\title{Initial maturity parameters for capelin example}
\format{
A data frame with 47 rows and 6 variables:
\describe{
  \item{year}{integer, year}
  \item{p1}{numeric, initial value for p1}
  \item{p2}{numeric, initial value for p2}
  \item{p3}{numeric, initial value for p3}
  \item{nu}{numeric, initial value for nu}
  \item{unknown}{integer, 0 or 1, ask Sam Subbey what this is}
}
}
\source{
\url{https://www.imr.no}
}
\usage{
maturityInitialParameters
}
\description{
A dataset containing p1,p2,p3 and nu initial parameters for running example
with min_age = 3 and max_age = 4 on capelin maturity.
}
\keyword{datasets}
