% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{plot.captoolSim}
\alias{plot.captoolSim}
\title{Plot captool projection (1Oct - 1Jan)}
\usage{
\method{plot}{captoolSim}(x, all = FALSE, ...)
}
\arguments{
\item{x}{simulation object}

\item{all}{logical, if TRUE all simulations are plotted with median in red, else only quantiles.}

\item{...}{additional arguments}
}
\value{
ggplot object
}
\description{
Plot captool projection (1Oct - 1Jan)
}
\examples{
\dontrun{
plot(sim)
plot(sim, all = TRUE)
}
}
