% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateConsumption.R
\name{plot.consumption}
\alias{plot.consumption}
\title{Plot consumption}
\usage{
\method{plot}{consumption}(x)
}
\arguments{
\item{x}{object of type consumption}
}
\value{

}
\description{
Plot consumption
}
\examples{
\dontrun{
 data("consumptionData")
par.list <- list(
  logCmax   = log(1.2),
  logChalf  = log(1e2),
  logalpha  = log(2),
  logbeta   = log(2),
  logSigma  = log(1e3)
)
cFit <- estimateConsumption(data = consumptionData, parameters = par.list, silent =TRUE,
                            map  = list(logalpha = factor(NA),
                                        logbeta = factor(NA)))
plot(cFit)
}
}
