% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaturity.R
\name{plot.maturitybyyears}
\alias{plot.maturitybyyears}
\alias{plot.maturity}
\alias{plot.summary.maturity}
\title{Plot maturity for all year classes}
\usage{
\method{plot}{maturitybyyears}(x, ...)

\method{plot}{maturity}(x, ...)

\method{plot}{summary.maturity}(x, type = "maturity", ...)
}
\arguments{
\item{x}{list, Result from runMaturityYearByYear}

\item{...}{additional arguments}

\item{type}{either "maturity" or "prediciton".}

\item{plot}{logical, indicating whether the figure is to be printed}
}
\value{
ggplot2 plot
}
\description{
Plot maturity for all year classes
}
\examples{
\dontrun{
data(cap)
data(catch)
data(maturityInitialParameters)
plot(runMaturityYearByYear(cap, catch=catch,
initPar =maturityInitialParameters[maturityInitialParameters$agegr =="2-3",],
 min_age = 2, max_age = 3, plot = FALSE))
 }
}
