% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMaturityYearByYear.R
\name{plot.maturitybyyears}
\alias{plot.maturitybyyears}
\title{Plot maturity by years}
\usage{
\method{plot}{maturitybyyears}(x, ...)
}
\arguments{
\item{x}{maturitybyyears}

\item{...}{}
}
\value{
ggplot object
}
\description{
Plot maturity by years
}
\examples{
data(cap)
data(catch)
data(maturityInitialParameters)
result <- runMaturityYearByYear(cap,catch,maturityInitialParameters, plot = TRUE)
plot.maturitybyyears(result)
}
