% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaturity.R
\name{plotMaturity}
\alias{plotMaturity}
\title{Plot maturity for all year classes}
\usage{
plotMaturity(x, plot = TRUE, ...)
}
\arguments{
\item{plot}{logical, indicating whether the figure is to be printed}

\item{result}{list, Result from runMaturityYearByYear}
}
\value{
ggplot2 plot
}
\description{
Plot maturity for all year classes
}
\examples{
#' data(cap)
data(catch)
data(maturityInitialParameters)
plotMaturity(runMaturityYearByYear(cap, catch=catch,
initPar =maturityInitialParameters, min_age = 3,
max_age = 4, plot = FALSE))
}
