% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifrost_search-methods.R
\name{print.bifrost_search}
\alias{print.bifrost_search}
\title{Print method for bifrost search results}
\usage{
\method{print}{bifrost_search}(x, ...)
}
\arguments{
\item{x}{A \code{bifrost_search} object returned by \code{\link[=searchOptimalConfiguration]{searchOptimalConfiguration()}}.}

\item{...}{Unused (S3 compatibility).}
}
\description{
Print method for bifrost search results
}
