% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateConsumption.R
\name{print.consumption}
\alias{print.consumption}
\title{Print consumption object}
\usage{
\method{print}{consumption}(x, ...)
}
\arguments{
\item{x}{consumption object}

\item{...}{additional arguments}
}
\value{
printout
}
\description{
Print consumption object
}
