% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaturity.R
\name{print.maturity}
\alias{print.maturity}
\title{Print maturity object}
\usage{
\method{print}{maturity}(x, ...)
}
\arguments{
\item{x}{maturity}

\item{...}{additional arguments}
}
\value{
printout
}
\description{
Print maturity object
}
