% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-map.R
\name{print.rateMap}
\alias{print.rateMap}
\title{Print a \code{rateMap} Object}
\usage{
\method{print}{rateMap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rateMap"} returned by \code{\link[=rateMap]{rateMap()}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{rateMap} Object
}
