% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateConsumption.R
\name{print.summary.consumption}
\alias{print.summary.consumption}
\title{Print summary of consumption}
\usage{
\method{print}{summary.consumption}(x, ...)
}
\arguments{
\item{x}{object of type summary.consumption}

\item{...}{addition arguments}
}
\value{
printout
}
\description{
Print summary of consumption
}
