% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaturity.R
\name{print.summary.maturity}
\alias{print.summary.maturity}
\title{Print summary of maturity}
\usage{
\method{print}{summary.maturity}(x, ...)
}
\arguments{
\item{x}{object of type summary.maturity}

\item{...}{addition arguments}
}
\value{
printout
}
\description{
Print summary of maturity
}
