% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captool_simulation.R
\name{q05ofcatch}
\alias{q05ofcatch}
\title{Function for calculating squared distance between 5% quantile and blim}
\usage{
q05ofcatch(
  totalcatch = 0,
  data_list,
  catch.distribution = c(0, 0.3, 0.7),
  blim = 0.2,
  seed = NULL,
  ...
)
}
\arguments{
\item{totalcatch}{total catch}

\item{data_list}{data_list as input for captool function}

\item{catch.distribution}{vector of length 3 given ratio of catch in jan, feb and mar}

\item{blim}{Blim}

\item{seed}{integer, for not changing the random input}

\item{...}{other arguments to be passed to captool function}
}
\description{
Function for calculating squared distance between 5% quantile and blim
}
