% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runFullSim}
\alias{runFullSim}
\title{Run full simulation from 1.October - 1.April}
\usage{
runFullSim(mFit, cFit, catches, nsim = 15000)
}
\arguments{
\item{mFit}{maturity fit object}

\item{cFit}{consumption fit object}

\item{catches}{catches data}

\item{nsim}{Number of simulations}
}
\value{
object of type captoolSim
}
\description{
Run full simulation from 1.October - 1.April
}
\examples{
\dontrun{
data(catch)
catches <- colSums(catch[catch$year == 2010, c("spring01", "spring05", "spring04")])
fSim <- runFullSim(mFit = mFit, cFit = cFit, cathces = catches, nsim = 15000)
plot(fSim)}
}
