% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMaturityYearByYear.R
\name{runMaturityYearByYear}
\alias{runMaturityYearByYear}
\title{Function for running maturity estimations year by year}
\usage{
runMaturityYearByYear(
  cap,
  catch,
  initPar,
  min_age = 3,
  max_age = 4,
  plot = TRUE
)
}
\arguments{
\item{cap}{data.frame, cap table. Should contain the following column names:
meanlength(cm), 1,2,3,4,5 and year, where 1-4 are ages and 5 is a 5+ group.}

\item{catch}{data.frame, catches data. Should contain 14 columns including
age, year and the remaining 12 for each month.}

\item{initPar}{data.frame, initial values for p1,p2,p3 and nu by year}

\item{min_age}{integer. Minimum age}

\item{max_age}{integer. Maximum age}

\item{plot}{logical, print plot (TRUE or FALSE)}
}
\value{
list, containing estimation output and results
}
\description{
Function for running maturity estimations year by year
}
\examples{
data(cap)
data(catch)
data(maturityInitialParameters)
result <- runMaturityYearByYear(cap,catch,maturityInitialParameters, plot = TRUE)
}
