% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaturity.R
\name{summary.maturity}
\alias{summary.maturity}
\title{Summary of estimation of Maturity}
\usage{
\method{summary}{maturity}(x)
}
\arguments{
\item{x}{maturity, object from running estimation}
}
\value{
summary
}
\description{
Summary of estimation of Maturity
}
