% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMaturityData.R
\name{updateMaturityData}
\alias{updateMaturityData}
\title{Update maturity data with new Nl object}
\usage{
updateMaturityData(data, Nl)
}
\arguments{
\item{data}{list, maturity data to be updated}

\item{Nl}{matrix, numbers at length group}
}
\value{
updated maturity data list
}
\description{
Update maturity data with new Nl object
}
