% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whichShifts}
\alias{whichShifts}
\title{Identify Nodes Representing Shifts in a SIMMAP Tree}
\usage{
whichShifts(tree)
}
\description{
This function identifies internal nodes in a SIMMAP-style phylogenetic tree
that represent regime shifts, based on the most recent common ancestors (MRCAs)
of tips sharing the same painted state.
}
