% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaturity.R
\name{estimateMaturity}
\alias{estimateMaturity}
\title{Function for estimating maturity}
\usage{
estimateMaturity(data, parameters, ...)
}
\arguments{
\item{data}{list of data inputs}

\item{parameters}{list of parameters}

\item{...}{additional arguments for TMB::MakeADFun}
}
\value{
list of TMB object and optimization object
}
\description{
Function for estimating maturity
}
\examples{
#estMature(data = data.list, parameters = par.list, silent =TRUE)
}
