% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractRegimeVCVs}
\alias{extractRegimeVCVs}
\title{Extract Regime-Specific Variance-Covariance Matrices from a BMM mvgls Model}
\usage{
extractRegimeVCVs(model_output)
}
\description{
This function extracts and scales the regime-specific variance-covariance (VCV) matrices
from a mvgls model fitted under the BMM (Brownian Motion with multiple regimes) framework.
It returns a named list of VCV matrices, each corresponding to an evolutionary regime.
}
