% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDescendants}
\alias{getDescendants}
\title{Get All Descendants of a Node in a Phylogenetic Tree}
\usage{
getDescendants(tree, node, include.node = FALSE)
}
\description{
Recursively retrieves all descendant nodes (internal and/or tips) from a specified node
in a phylogenetic tree of class \code{"phylo"}. Optionally includes the starting node in the output.
}
