% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDescendants}
\alias{getDescendants}
\title{Get All Descendants of a Node in a Phylogenetic Tree}
\usage{
getDescendants(tree, node, include.node = FALSE)
}
\arguments{
\item{tree}{An object of class \code{"phylo"} (see \code{\link[ape]{phylo}}).}

\item{node}{An integer specifying the node number from which to retrieve descendants.
Internal nodes are numbered from \code{Ntip(tree) + 1} to
\code{Ntip(tree) + Nnode(tree)} in the \code{phylo} object.}

\item{include.node}{Logical (default \code{FALSE}). If \code{TRUE}, the specified
node itself is included in the returned vector.}
}
\value{
An integer vector of node numbers (tips and/or internal nodes)
that descend from the specified node. The order of descendants is not guaranteed.
}
\description{
Recursively retrieves all descendant nodes (internal nodes and tips) from a specified
node in a phylogenetic tree of class \code{"phylo"}.
}
\details{
This function performs a depth-first traversal of the tree, recursively
identifying all nodes downstream of the specified starting node.
}
\examples{
library(ape)
set.seed(123)
tree <- rtree(5)

# Get descendants of the root node
root_node <- Ntip(tree) + 1
getDescendants(tree, root_node)

# Include the starting node itself
getDescendants(tree, root_node, include.node = TRUE)

}
\seealso{
\code{\link[ape]{Ntip}}, \code{\link[ape]{Nnode}},
\code{\link[ape]{getMRCA}}, \code{\link[phytools]{getParent}}
}
