% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paintSubTree_removeShift}
\alias{paintSubTree_removeShift}
\title{Remove a Painted Shift from a Subtree in a SIMMAP-Formatted Phylogenetic Tree}
\usage{
paintSubTree_removeShift(tree, shift_node, stem = FALSE)
}
\description{
This function removes a previously painted shift (regime change) from a specified node and its descendant branches
in a SIMMAP-style phylogenetic tree by selectively overwriting branches painted with the shift state.
The original ancestral state (inherited from the parent node) is restored, with optional stem painting of the edge leading to the node.
}
