% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-map.R
\name{plotRateMap}
\alias{plotRateMap}
\alias{plot.rateMap}
\title{Plot a \code{rateMap} Object}
\usage{
plotRateMap(x, ...)

\method{plot}{rateMap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rateMap"} returned by \code{\link[=rateMap]{rateMap()}}.}

\item{...}{Additional plotting controls. Supported options include
\code{legend}, \code{fsize}, \code{tip_fsize}, \code{legend_fsize}, \code{ftype},
\code{show_tip_labels}, \code{outline}, \code{lwd}, \code{type}, \code{mar}, \code{direction},
\code{offset}, \code{xlim}, \code{ylim}, \code{hold}, \code{underscore}, and \code{arc_height}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Render a rate-variation map produced by \code{\link[=rateMap]{rateMap()}} using
\code{\link[phytools:plotSimmap]{phytools::plotSimmap()}} and a continuous color-bar legend.
}
\examples{
\dontrun{
rm_obj <- rateMap(fits, plot = FALSE, progress = FALSE)
plotRateMap(
  rm_obj,
  type = "arc",
  show_tip_labels = FALSE,
  legend_fsize = 0.8
)
}

}
\seealso{
\code{\link[=rateMap]{rateMap()}}
}
