% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captool_simulation.R
\name{plotting_validation}
\alias{plotting_validation}
\title{Plotting output from captool}
\usage{
plotting_validation(return, path = "", save = FALSE)
}
\arguments{
\item{return}{list, returned by captool function}

\item{path}{character, where to save the figure}

\item{save}{boolean, should the figure be saved?}
}
\value{
ggplot figure object
}
\description{
Plotting output from captool
}
