% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchOptimalConfiguration.R
\name{searchOptimalConfiguration}
\alias{searchOptimalConfiguration}
\title{Search for the Optimal Shift Configuration in a Phylogenetic Model}
\usage{
searchOptimalConfiguration(
  baseline_tree,
  trait_data,
  formula = "trait_data~1",
  min_descendant_tips,
  num_cores = 2,
  ic_uncertainty_threshold = 1,
  shift_acceptance_threshold = 1,
  uncertainty = F,
  uncertaintyweights = F,
  uncertaintyweights_par = F,
  postorder_traversal = F,
  plot = T,
  IC = "GIC",
  store_model_fit_history = TRUE,
  ...
)
}
\description{
This function performs a stepwise search to identify the optimal configuration of evolutionary regime shifts
on a phylogenetic tree, using information criteria (GIC or BIC) to guide model selection. It evaluates candidate
shifts in parallel, fits models at each stage, optionally accounts for model uncertainty, and computes model weights.
}
