% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateConsumption.R
\name{summary.consumption}
\alias{summary.consumption}
\title{Summary of estimation of Consumption}
\usage{
\method{summary}{consumption}(obj, ...)
}
\arguments{
\item{obj}{consumption, object from running estimation}

\item{...}{additional arguments}
}
\value{
summary
}
\description{
Summary of estimation of Consumption
}
