% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMaturity.R
\name{summary.maturity}
\alias{summary.maturity}
\title{Summary of estimation of Maturity}
\usage{
\method{summary}{maturity}(obj, ...)
}
\arguments{
\item{obj}{maturity, object from running estimation}

\item{...}{additional arguments}
}
\value{
summary
}
\description{
Summary of estimation of Maturity
}
